@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.3 <4.0.0'
+  sdk: '>=3.4.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,16 +34,17 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  flutter_smart_dialog: ^4.9.5+1
-  get: ^4.6.5
+  cupertino_icons: ^1.0.6
+  flutter_smart_dialog: ^4.9.7+8
+  get: ^4.6.6
   desktop_drop: ^0.4.4
-  cross_file: ^0.3.3+6
-  file_picker: ^6.1.1
-  path_provider: ^2.1.1
+  cross_file: ^0.3.4+1
+  file_picker: ^8.0.5
+  path_provider: ^2.1.3
   path: ^1.8.3
-  shared_preferences: ^2.2.2
+  shared_preferences: ^2.2.3
   qr_flutter: ^4.1.0
+  win32: 5.8.0
 
 dev_dependencies:
   flutter_test:
